+++ source/index.html
@@ -77,13 +77,12 @@
   </header>
 
   <main class="page-content">
-    <canvas id="lessonCanvas"></canvas>
-
     <section class="animation-screen"></section>
 
     <section class="screen screen--hidden screen--intro js-scrollspy" id="top">
       <div class="screen__wrapper">
         <div class="intro">
+          <canvas id="introCanvas"></canvas>
           <h1 class="intro__title">Таинственный отпуск</h1>
           <div class="intro__content">
             <div class="intro__info">
