+++ source/js/helpers/canvas-frame.js
