+++ source/scss/blocks/animation-screen.scss
@@ -5,20 +5,20 @@
   align-items: center;
   width: 100%;
   height: 100%;
-  background-image: url("../img/scene.png");
+  // background-image: url("../img/scene.png");
   background-repeat: no-repeat;
   background-size: 100% auto;
   background-position: center;
 
-  @include retina {
-    background-image: url("../img/scene@2x.png");
-  }
+  // @include retina {
+  //   background-image: url("../img/scene@2x.png");
+  // }
 
-  @media (orientation: portrait) {
-    background-image: url("../img/scene-mob.png");
+  // @media (orientation: portrait) {
+  //   background-image: url("../img/scene-mob.png");
 
-    @include retina {
-      background-image: url("../img/scene-mob@2x.png");
-    }
-  }
+  //   @include retina {
+  //     background-image: url("../img/scene-mob@2x.png");
+  //   }
+  // }
 }
