+++ source/scss/style.scss
@@ -20,10 +20,3 @@
 @import "blocks/screen.scss";
 @import "blocks/slider.scss";
 @import "blocks/social-block.scss";
-
-#lessonCanvas {
-  position: fixed;
-  z-index: 1000000000;
-  top: 0;
-  left: 0;
-}
